Moved aeson instances of OsPath to OsPath.Aeson namespace
