+++ path.cabal
@@ -15,6 +15,7 @@ extra-source-files:  README.md
                    , CHANGELOG
                    , src/Path/Include.hs
                    , src/Path/Internal/Include.hs
+                   , src/OsPath/Aeson/Include.hs
                    , src/OsPath/Include.hs
                    , src/OsPath/Internal/Include.hs
                    , test/Common/Include.hs
@@ -36,6 +37,9 @@ library
                    , Path.Internal.Posix
                    , Path.Internal.Windows
                    , OsPath
+                   , OsPath.Aeson
+                   , OsPath.Aeson.Posix
+                   , OsPath.Aeson.Windows
                    , OsPath.Posix
                    , OsPath.Windows
                    , OsPath.Internal
