+++ src/OsPath/Aeson/Windows.hs
@@ -2,6 +2,5 @@
 #define PLATFORM_NAME        Windows
 #define PLATFORM_PATH        WindowsPath
 #define PLATFORM_PATH_SINGLE 'WindowsPath'
-#define PLATFORM_STRING      WindowsString
-#define IS_WINDOWS           1
+#define PLATFORM_UTF_CODEC   UTF16-LE
 #include "Include.hs"
